Add Error Middleware, &  Token Generation Helper
Every Route Handler will be treated as a middleware, if a particular
handler throws an error, the error will be captured in the last
middleware.

We make this middleware to handle all errors that will occur within our
API gracefully.

To generate a token when the user successfully logs in, the
generateToken function helps us in doing so